Moved ranks and totals inside
